+++ desktop/cmp/grouping/GroupingChooser.ts
@@ -5,8 +5,9 @@
  * Copyright © 2023 Extremely Heavy Industries Inc.
  */
 import {GroupingChooserModel} from '@xh/hoist/cmp/grouping';
+import {GroupingChooserLocalModel} from '@xh/hoist/cmp/grouping/impl/GroupingChooserLocalModel';
 import {box, div, filler, fragment, hbox, vbox} from '@xh/hoist/cmp/layout';
-import {hoistCmp, HoistModel, HoistProps, useLocalModel, uses} from '@xh/hoist/core';
+import {hoistCmp, HoistProps, useLocalModel, uses} from '@xh/hoist/core';
 import {button, ButtonProps} from '@xh/hoist/desktop/cmp/button';
 import {MENU_PORTAL_ID, select} from '@xh/hoist/desktop/cmp/input';
 import {panel} from '@xh/hoist/desktop/cmp/panel';
@@ -14,10 +15,9 @@ import '@xh/hoist/desktop/register';
 import {Icon} from '@xh/hoist/icon';
 import {menu, menuDivider, menuItem, popover} from '@xh/hoist/kit/blueprint';
 import {dragDropContext, draggable, droppable} from '@xh/hoist/kit/react-beautiful-dnd';
-import {action, computed, makeObservable, observable} from '@xh/hoist/mobx';
-import {createObservableRef, splitLayoutProps} from '@xh/hoist/utils/react';
+import {splitLayoutProps} from '@xh/hoist/utils/react';
 import classNames from 'classnames';
-import {compact, difference, isEmpty, isEqual, sortBy} from 'lodash';
+import {compact, isEmpty, sortBy} from 'lodash';
 import './GroupingChooser.scss';
 
 export interface GroupingChooserProps extends ButtonProps<GroupingChooserModel> {
@@ -38,126 +38,6 @@ export interface GroupingChooserProps extends ButtonProps<GroupingChooserModel> 
 
     /** True (default) to style target button as an input field - blends better in toolbars. */
     styleButtonAsInput?: boolean;
-}
-
-class GroupingChooserLocalModel extends HoistModel {
-    private readonly model: GroupingChooserModel;
-
-    @observable.ref pendingValue: string[] = [];
-    @observable editorIsOpen: boolean = false;
-    @observable favoritesIsOpen: boolean = false;
-
-    popoverRef = createObservableRef<HTMLElement>();
-
-    constructor(model: GroupingChooserModel) {
-        super();
-        makeObservable(this);
-
-        this.model = model;
-
-        this.addReaction({
-            track: () => this.pendingValue,
-            run: () => {
-                if (model.commitOnChange) model.setValue(this.pendingValue);
-            }
-        });
-    }
-
-    @computed
-    get availableDims(): string[] {
-        return difference(this.model.dimensionNames, this.pendingValue);
-    }
-
-    @computed
-    get isValid(): boolean {
-        return this.model.validateValue(this.pendingValue);
-    }
-
-    @computed
-    get isAddEnabled(): boolean {
-        const {pendingValue, availableDims} = this,
-            {maxDepth, dimensionNames} = this.model,
-            limit =
-                maxDepth > 0 ? Math.min(maxDepth, dimensionNames.length) : dimensionNames.length,
-            atMaxDepth = pendingValue.length === limit;
-        return !atMaxDepth && !isEmpty(availableDims);
-    }
-
-    @action
-    toggleEditor() {
-        this.pendingValue = this.model.value;
-        this.editorIsOpen = !this.editorIsOpen;
-        this.favoritesIsOpen = false;
-    }
-
-    @action
-    toggleFavoritesMenu() {
-        this.favoritesIsOpen = !this.favoritesIsOpen;
-        this.editorIsOpen = false;
-    }
-
-    @action
-    closePopover() {
-        this.editorIsOpen = false;
-        this.favoritesIsOpen = false;
-    }
-
-    //-------------------------
-    // Value handling
-    //-------------------------
-
-    @action
-    addPendingDim(dimName: string) {
-        if (!dimName) return;
-        this.pendingValue = [...this.pendingValue, dimName];
-    }
-
-    @action
-    replacePendingDimAtIdx(dimName: string, idx: number) {
-        if (!dimName) return this.removePendingDimAtIdx(idx);
-        const pendingValue = [...this.pendingValue];
-        pendingValue[idx] = dimName;
-        this.pendingValue = pendingValue;
-    }
-
-    @action
-    removePendingDimAtIdx(idx: number) {
-        const pendingValue = [...this.pendingValue];
-        pendingValue.splice(idx, 1);
-        this.pendingValue = pendingValue;
-    }
-
-    @action
-    movePendingDimToIndex(dimName: string, toIdx: number) {
-        const pendingValue = [...this.pendingValue],
-            dim = pendingValue.find(it => it === dimName),
-            fromIdx = pendingValue.indexOf(dim);
-
-        pendingValue.splice(toIdx, 0, pendingValue.splice(fromIdx, 1)[0]);
-        this.pendingValue = pendingValue;
-    }
-
-    @action
-    commitPendingValueAndClose() {
-        const {pendingValue, model} = this,
-            {value} = model;
-
-        if (!isEqual(value, pendingValue) && model.validateValue(pendingValue)) {
-            model.setValue(pendingValue);
-        }
-
-        this.closePopover();
-    }
-
-    //--------------------
-    // Drag Drop
-    //--------------------
-
-    onDragEnd(result) {
-        const {draggableId, destination} = result;
-        if (!destination) return;
-        this.movePendingDimToIndex(draggableId, destination.index);
-    }
 }
 
 /**
